pkg/backend/pyproject.toml
[project]
name = "caze-bizcon-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "3.12.10"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-core (>=0.3.65,<0.4.0)",
    "langchain-chroma (>=0.2.4,<0.3.0)",
    "langchain-openai (>=0.3.23,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "cryptography (>=45.0.4,<46.0.0)",
    "pystray (>=0.19.5,<0.20.0)",
    "google-auth (>=2.0)",
    "google-auth-oauthlib (>=1.0)",
    "google-api-python-client (>=2.0)",
    "pymupdf4llm (>=0.0.25,<0.0.26)",
    "python-docx (>=1.1.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
